# Integrate views with mixins