--- a/beluga/include/beluga/mixin.hpp
+++ b/beluga/include/beluga/mixin.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Ekumen, Inc.
+// Copyright 2023-2024 Ekumen, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,6 +18,9 @@
 #include <memory>
 
 #include <beluga/mixin/descriptor.hpp>
+#include <beluga/mixin/particle_filter.hpp>
+#include <beluga/mixin/sampling.hpp>
+#include <beluga/mixin/storage.hpp>
 #include <beluga/mixin/utility.hpp>
 #include <ciabatta/ciabatta.hpp>