--- a/beluga/test/beluga/mixin/test_storage.cpp
+++ b/beluga/test/beluga/mixin/test_storage.cpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Ekumen, Inc.
+// Copyright 2023-2024 Ekumen, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 
 #include <gmock/gmock.h>
 
-#include <beluga/storage.hpp>
+#include <beluga/mixin/storage.hpp>
 #include <ciabatta/ciabatta.hpp>
 
 #include <range/v3/algorithm/equal.hpp>
@@ -78,7 +78,7 @@ TYPED_TEST(StoragePolicyTest, ResampleParticles) {
   EXPECT_CALL(mixin, max_samples()).WillOnce(Return(5)).WillOnce(Return(5));
   mixin.initialize_states(states);
   ASSERT_EQ(mixin.particle_count(), 5);
-  mixin.initialize_particles(states | ranges::views::reverse);
+  mixin.initialize_particles(mixin.particles() | ranges::views::reverse);
   ASSERT_TRUE(ranges::equal(mixin.states(), states | ranges::views::reverse));
 }